+++ .github/actions/storage/restore/main.ts
@@ -4,9 +4,14 @@ import {setTimeout} from 'node:timers/promises'
 import {restoreCache} from '@actions/cache'
 import * as core from '@actions/core'
 
-if (process.platform === 'linux' && existsSync('/etc/alpine-release')) {
-  core.debug('alpine system is detected, installing necessary dependencies')
-  execSync('apk add --no-cache zstd tar')
+if (process.platform === 'linux') {
+  if (existsSync('/etc/alpine-release')) {
+    core.debug('alpine system is detected, installing necessary dependencies')
+    execSync('apk add --no-cache zstd tar')
+  } else if (execSync('cat /etc/*release | grep ^ID=', {encoding: 'utf-8'}).includes('debian')) {
+    core.debug('debian system is detected, installing necessary dependencies')
+    execSync('apt-get update && apt-get install -y zstd')
+  }
 }
 
 main()
@@ -26,20 +31,28 @@ async function main(): Promise<(string | undefined)[]> {
     const [name, paths] = compositeName.split('$')
     const fallbacks = latest ? [name.replace(/(?<=#).+$/, '')] : []
 
-    return restore({paths: paths.split(';'), name, fallbacks, wait})
+    return restore({paths: paths.split(';'), name, fallbacks, wait: wait ? 600_000 : 0})
   }))
 
-  async function restore(options: {paths: string[], name: string, fallbacks: string[], wait?: boolean}): Promise<string | undefined> {
-    // NOTE: restoreCache mutates paths argument, that makes it impossible to reuse
-    const paths = [...options.paths]
-    const restoredName = await restoreCache(options.paths, options.name, options.fallbacks, {}, true)
+  async function restore(options: {
+    paths: string[],
+    name: string,
+    fallbacks: string[],
+    wait?: number,
+    startedAt?: number
+  }): Promise<string | undefined> {
+    options.startedAt ??= Date.now()
+    const restoredName = await restoreCache([...options.paths], options.name, options.fallbacks, {}, true)
     if (restoredName) {
       core.info(`cache was successfully restored with ${options.name}`)
       return restoredName
-    } else if (wait) {
+    } else if (options.wait) {
+      if (options.startedAt + options.wait <= Date.now()) {
+        throw new Error(`Failed to restore artifact during ${options.wait} ms`)
+      }
       core.info(`waiting for cache with name ${options.name} to appear`)
       await setTimeout(20_000)
-      return restore({...options, paths})
+      return restore(options)
     }
   }
 }