+++ js/packages/eyes-cypress/test/fixtures/testApp/cypress/integration-run/retries.js
@@ -3,7 +3,7 @@ describe('remove duplicate tests on retry', () => {
   it('fails and retries', {retries: 2}, () => {
     const retryCount = cy.state('runnable')._currentRetry
     cy.visit('https://example.org').then(() => {
-      // we wait here to make sure we get different time stamps for all tests
+      // we wait here to make sure we get different time stamps for all tests 
       cy.wait(1000)
       if (retryCount < 2) cy.document().then(doc => (doc.body.style.background = 'red'))
     })