+++ js/packages/ufg-client/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [1.7.0](https://github.com/applitools/eyes.sdk.javascript1/compare/js/ufg-client@1.6.0...js/ufg-client@1.7.0) (2023-08-08)
+
+
+### Features
+
+* allow providing custom headers for resource fetching  ([#1852](https://github.com/applitools/eyes.sdk.javascript1/issues/1852)) ([372cb96](https://github.com/applitools/eyes.sdk.javascript1/commit/372cb96b905a0661c36e2fa10a7855208fb55bb0))
+
+## [1.6.0](https://github.com/applitools/eyes.sdk.javascript1/compare/js/ufg-client@1.5.3...js/ufg-client@1.6.0) (2023-08-03)
+
+
+### Features
+
+* add stuck request retries to all requests to UFG and Eyes ([#1826](https://github.com/applitools/eyes.sdk.javascript1/issues/1826)) ([5884d42](https://github.com/applitools/eyes.sdk.javascript1/commit/5884d428b230e3a832a2110a388ebe63a94006fc))
+
+
+### Bug Fixes
+
+* improve fetch error experience when fetching from tunnel ([e7d8b49](https://github.com/applitools/eyes.sdk.javascript1/commit/e7d8b49947c07beb27f110cb4952e8c3206469af))
+
 ## [1.5.3](https://github.com/applitools/eyes.sdk.javascript1/compare/js/ufg-client@1.5.2...js/ufg-client@1.5.3) (2023-07-27)